power: supply: qti_battery_charger: Return charge counter in uAh
Previously the charge counter was not reported in uAh while the
charge full capacity was. The charge counter used to go down to
1000 times as low as the max battery capacity is in uAh.
Frameworks use this property to set the battery state and
expects it to be in uAh. Refer to [1] and [2].

Since commit "Use getUahDischarge when available" [3] AOSP
uses the getUahDischarge method to determine the amount of
discharge. Without this patch the value used to be 1000 times
too low leading to the consumed power of
AGGREGATE_BATTERY_CONSUMER_SCOPE_DEVICE becoming very low
so that the relative consumption of all consumers got very
high.

[1]: https://github.com/LineageOS/android_frameworks_base/blob/d42ec4c0e5ad85792f526cc90d8830b02febc1b0/services/core/java/com/android/server/BatteryService.java#L579
[2]: https://github.com/LineageOS/android_frameworks_base/blob/d42ec4c0e5ad85792f526cc90d8830b02febc1b0/services/core/java/com/android/server/am/BatteryStatsService.java#L2079
[3]: LineageOS/android_frameworks_base@9f08c60

Change-Id: I64dff6478588d991971b60bad61652f5c9f581b0
Picked from: https://github.com/cupid-development/android_kernel_xiaomi_sm8450/commit/f5d4982335a28dfbc9bc29a6f255e345dd72faa6

Signed-off-by: Blinkov `rifux` Vladimir ♪ <87264036+rifux@users.noreply.github.com>